requirements.txt
cv2
Pillow
numpy